Add loading spinners, convert code to async (#20)
This cleans up the output of the command. It was very noisy before, and
it wasn't obvious which commands the output was related to (esp. when
running `create-expo-app`). We might want to still preserve command
output and write it to a log file.

Also, we were previously using a mixture of sync and async code. This PR
updates the codebase to be fully asynchronous, since the loading spinner
cannot animate when using sync code.


https://github.com/thoughtbot/thoughtbelt/assets/1691324/0b5bfec2-1d01-476d-ac0b-7cc15c2e72e5