Create App.tsx at end of "createApp" command, since will depend on other tasks (#25)
When we create a new app using thoughtbelt, we will want to generate an
`App.tsx` that has all of our providers and tools configured, including
React Navigation, global state provider, styling provider, etc. To me,
since this file will be tied to many different commands, it would make
more sense for the creation of this file to be a step at the end of the
`createApp` command, independent of any one command.

If we run `thoughtbelt navigation` independently, we will not want to
create a new `App.tsx` in that case, though it would likely be helpful
to include some output directing the user to wrap their app in a
`NavigationContainer`. What do you think?

We don't yet have a mechanism for detecting if a given command is
running independently as opposed to being a part of the main `createApp`
(or other) command .