Fix async-storage command (#27)
This PR installs `AsyncStorage` and `react-native-safe-area-context`
with `npx expo install` instead of a regular `npm install`. The Expo
docs recommend using `npx expo install` for these packages, which
ensures that the correct versions are added that have been tested with
the current Expo version. We seeing a warning that we were using an
incompatible version of AsyncStorage.

Additionally, this PR adds `thoughtbelt` as a dev dependency in the new
project so users can continue to use thoughtbelt with eg. `yarn belt`.