Refactor copyTemplateDirectory
Previously, the flow was:

1. Copy the template directory to the destination directory
2. Iterate through each of the files copied
  1. Perform string substitutions
  2. Render ETA templates (and remove .eta)

When copying the "boilerplate" directory, we ideally ignore any files
that are "gitignored" by this repo. This can be especially helpful with
the "boilerplate" directory in the case where a developer has run the
app and has a node_modules directory there.

The new approach is:

1. Iterate through each of the files in the template directory
2. Filter out files that are gitignored (based on gitignore contents
   passed in)
3. For each file, read contents, perform substitutions, copy to
   destination