Update createApp to no longer use individual commands (#33)
This PR addresses tickets: [Architecture of "create" vs individual
commands](https://trello.com/c/KNNHwx2Q/45-architecture-of-create-vs-individual-commands)
and [How to prevent flakiness when versions
change?](https://trello.com/c/lzS7nr7R/44-how-to-prevent-flakiness-when-versions-change).

I recommend reading the context from those tickets for background
information. The issues this PR seeks to address are:

## Create architecture

### Issues

The `create` command currently runs a collection of isolated individual
commands. These commands sometimes modify similar files, like
`package.json`, `App.tsx`, test files, etc. Because these commands touch
the same files, it becomes difficult to "layer" these changes on top of
each other. Each command is unable to make assumptions about the state
of the file and must instead have a way to mutate the existing file. A
secondary issue is that it becomes difficult to reason about what the
final output of the `create` command is without actually creating a new
app (no easy way to visualize how the individual commands layer
together).

### Solution

The `create` command has been refactored to no longer run any other
commands. Instead, we have a `boilerplate` directory that we copy
directly to the new app directory. After that, we perform a few string
substitutions to insert the app name in the appropriate places
(`package.json`, `app.json`) and then install dependencies with the
specified package manager. That's it!

The `templates/boilerplate` directory is the entire app! You can
literally `cd templates/boilerplate` then run `npm install` and `npm run
ios` or `npm run test:all` to run the app straight from there while
updating it! This is a huge improvement in the dev experience and
greatly simplifies understanding of what we're generating. This PR does
not introduce any changes in the app that gets generated -- it's
identical to what it was before. I generated the `boilerplate` directory
in this PR by creating an app using our old method and then copying the
result to `boilerplate`.

## Preventing flakiness

### Issues

Previously, Belt has been using the latest version of Expo, and the
latest version of any other dependencies it installs. Always being on
the latest and greatest is certainly desired, but it comes at the cost
of stability. If any dependency updates to a new major or minor version,
there is a chance that Belt might quit working. This actually happened
when Expo 50 was released (see #31).

### Solution

We need to balance being up to date with being stable. While we want
both, stability should be a non-negotiable, so let's prioritize
stability and then find ways to simplify the dependency upgrade process
so we can stay as up-to-date as possible. If we can streamline updating
Belt to support newer versions of Expo or other dependencies to take
just a few minutes, then we are more likely to be able to keep
up-to-date.