Add validator for the application name
Before, when a user introduce an application name that includes spaces,
the script will exit, throwing an error.

In this PR, we confirm the application name by checking that the:

It does include something other than numbers.

It does not match the string belt, as it is the name of our package.

If it includes spaces, we parse the name and return it with the first
letter of each word upper cased and joined without spaces.