Schedule CI workflow to run every night (#41)
In a GitHub action we can add the
[schedule](https://docs.github.com/en/actions/using-workflows/events-that-trigger-workflows#schedule)
event to run a workflow at a specific time.
In this case we will run the workflow every day at midnight UTC.

The notifications about the workflow will be sent to the user who set it
up, in this case myself. More information below.
> Notifications for scheduled workflows are sent to the user who
initially created the workflow. If a different user updates the cron
syntax in the workflow file, subsequent notifications will be sent to
that user instead. If a scheduled workflow is disabled and then
re-enabled, notifications will be sent to the user who re-enabled the
workflow rather than the user who last modified the cron syntax.

[Information about
notifications.](https://docs.github.com/en/actions/monitoring-and-troubleshooting-workflows/notifications-for-workflow-runs)

---------

Co-authored-by: Stefanni Brasil <stefanni@thoughtbot.com>