Pin '@types/react' to 18.2 (#43)
I was seeing this warning when running `npm run ios` on a newly
installed app:

> The following packages should be updated for best compatibility with
the installed expo version:
>    @types/react@18.3.0 - expected version: ~18.2.45
> Your project may not work correctly until you install the correct
versions of the packages.

I noticed that we were only pinning the major version of `@types/react`,
so I updated to also pin the minor version to match the React minor
version 18.2.x. Now, I do not see that warning when starting a new Belt
app.